Refactor dir structure and filenames
