+++ contracts/upgradeable_contracts/BasicOmnibridge.sol
@@ -1,9 +1,9 @@
 pragma solidity 0.7.5;
 
 import "@openzeppelin/contracts/math/SafeMath.sol";
-import "../Initializable.sol";
-import "../Upgradeable.sol";
-import "../Claimable.sol";
+import "./Initializable.sol";
+import "./Upgradeable.sol";
+import "./Claimable.sol";
 import "./components/bridged/BridgedTokensRegistry.sol";
 import "./components/native/NativeTokensRegistry.sol";
 import "./components/native/MediatorBalanceStorage.sol";
@@ -12,14 +12,14 @@ import "./components/common/OmnibridgeInfo.sol";
 import "./components/common/TokensBridgeLimits.sol";
 import "./components/common/FailedMessagesProcessor.sol";
 import "./modules/factory/TokenFactoryConnector.sol";
-import "../../interfaces/IBurnableMintableERC677Token.sol";
-import "../../libraries/TokenReader.sol";
+import "../interfaces/IBurnableMintableERC677Token.sol";
+import "../libraries/TokenReader.sol";
 
 /**
- * @title BasicMultiAMBErc20ToErc677
- * @dev Common functionality for universal multi-erc20-to-erc677 mediator intended to work on top of AMB bridge.
+ * @title BasicOmnibridge
+ * @dev Common functionality for multi-token mediator intended to work on top of AMB bridge.
  */
-abstract contract BasicMultiAMBErc20ToErc677 is
+abstract contract BasicOmnibridge is
     Initializable,
     Upgradeable,
     Claimable,
