+++ deploy/.env.example
@@ -1,4 +1,4 @@
-BRIDGE_MODE=MULTI_AMB_ERC20_TO_ERC677
+BRIDGE_MODE=OMNIBRIDGE
 
 DEPLOYMENT_ACCOUNT_PRIVATE_KEY=67..14
 DEPLOYMENT_GAS_LIMIT_EXTRA=0.2
