+++ deploy/src/loadContracts.js
@@ -1,7 +1,7 @@
 module.exports = {
   EternalStorageProxy: require(`../../build/contracts/EternalStorageProxy.json`),
-  HomeMultiAMBErc20ToErc677: require(`../../build/contracts/HomeMultiAMBErc20ToErc677.json`),
-  ForeignMultiAMBErc20ToErc677: require(`../../build/contracts/ForeignMultiAMBErc20ToErc677.json`),
+  HomeOmnibridge: require(`../../build/contracts/HomeOmnibridge.json`),
+  ForeignOmnibridge: require(`../../build/contracts/ForeignOmnibridge.json`),
   TokenFactory: require(`../../build/contracts/TokenFactory.json`),
   MultiTokenForwardingRulesManager: require(`../../build/contracts/MultiTokenForwardingRulesManager.json`),
   PermittableToken: require(`../../precompiled/PermittableToken.json`),
